pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "This is a google gemeni Ai assistant made by Haider Hussain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "package-name>=0.1",
    "python-dotenv>=1.1.0",
]